fix: align SQL schemas and README docs with Java types
- Updated H2, PostgreSQL, and Oracle create table scripts to use BIGINT (or NUMBER(19)) for EXPIRE_SEC to align with the Java API's `long expirationSeconds`.
- Updated H2 and Oracle scripts to use TIMESTAMP instead of DATETIME for consistency.
- Updated main README.md SQL example to use TIMESTAMP and BIGINT.
- Updated dlock-api/README.md to exactly match the Javadocs in KeyLock.java, including `{@value}` references and updated tags.

+++ dlock-jdbc/src/main/resources/db/ORACLE-create.sql
@@ -2,6 +2,6 @@ CREATE TABLE "@@tableName@@" (
   "LCK_KEY" VARCHAR2(1000) PRIMARY KEY,
   "LCK_HNDL_ID" VARCHAR2(100) not null,
   "CREATED_TIME" TIMESTAMP not null,
-  "EXPIRE_SEC" INTEGER not null
+  "EXPIRE_SEC" NUMBER(19) not null
 );
 CREATE UNIQUE INDEX "@@tableName@@_HNDL_UX" ON "@@tableName@@" ("LCK_HNDL_ID");
